Add some more stuff, fix subtraction bug

+++ play.h
@@ -4,8 +4,8 @@
 
 #include "konane.h"
 
-int64_t negamax(const State board[SIZE][SIZE], const State color, const uint32_t depth);
-int64_t computer_move(Move* move, const State board[SIZE][SIZE], const State player);
+int32_t negamax(const State board[SIZE][SIZE], const State color, const uint32_t depth);
+int32_t computer_move(Move* move, const State board[SIZE][SIZE], const State player);
 
 static inline void board_copy(State dst[SIZE][SIZE], const State src[SIZE][SIZE]) {
     for (size_t i = 0; i < SIZE; ++i) {
